- Move SubSystem class to Core - Add GetDataID to Kinematics interface - Doxygen: enable JAVADOC_AUTOBRIEF
